% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{fit_summary}
\alias{fit_summary}
\title{Title}
\usage{
fit_summary(my_data, example, fitted_values, m, response, beta_i = 0)
}
\arguments{
\item{my_data}{my_data the data set you're using}

\item{example}{the name of your example: will be one of either "culcita", "ctsib", "epilepsy", or "tortoise"}

\item{fitted_values}{fitted values of your linear predictor under the null hypothesis without random effects}

\item{m}{either the number of repeated observations from n individuals (repeated measures study)
or the number of subjects within n groups (grouped effects data)}

\item{response}{the response variable of your data set in ""}

\item{beta_i}{the parameter of interest for which we wish to test the null hypothesis for, default value of 0 for the intercept}
}
\value{
a list with a tibble of the estimates, standard errors, and confidence intervals and a p-value for the beta under the null hypothesis
}
\description{
Title
}
\examples{
load("~/Gopher.Fit/data/tortoise.rda")

#Fit model to gopher tortoises data
tortoise_fit <- run_model(data = tortoise, example = "tortoise")

# Obtain the fitted values
x1 <- tortoise$prev
x2 <- tortoise$Area

fitted_values <- tortoise_fit$beta[[1]] + tortoise_fit$beta[[2]] * x1 + log(x2) #fitted values without random effects

# Find summary
fit_summary(tortoise, "tortoise", fitted_values, 3, "shells", "prev")
}
