% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gopher.Fit-package.R
\docType{package}
\name{Gopher.Fit-package}
\alias{Gopher.Fit}
\alias{Gopher.Fit-package}
\title{Gopher.Fit: Fit Poisson GLMM}
\description{
This package aims to fit a Poisson GLMM. The package provides functions to fit the model and provide summary information including the point estimates, standard errors, and a 95% confidence interval for each parameter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesFB1/Gopher.Fit}
  \item Report bugs at \url{https://github.com/JamesFB1/Gopher.Fit/issues}
}

}
\author{
\strong{Maintainer}: James Fraser Barsby \email{jfraserb@uwo.ca}

Authors:
\itemize{
  \item Shi Yu (Dora) Cao \email{scao85@uwo.ca}
}

}
\keyword{internal}
