% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_hypothesis_test.R
\name{Hypothesis_Test}
\alias{Hypothesis_Test}
\title{Final Project Hypothesis Test}
\usage{
Hypothesis_Test(my_data = tortoise)
}
\arguments{
\item{my_data}{the data set you're using}
}
\value{
the p-value obtained after testing then null hypothesis
}
\description{
Final Project Hypothesis Test
}
\examples{
# load tortoise data
load("~/Gopher.Fit/data/tortoise.rda")

# Compute p-value
Hypothesis_Test(my_data = tortoise)

}
