% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tortoise}
\alias{tortoise}
\title{Gopher Tortoise data}
\format{
A data frame with 30 rows and 7 variables:
\describe{
\item{Site}{Site name}
\item{year}{year}
\item{shells}{number of shells found in site}
\item{type}{type of tortoise}
\item{Area}{area of site}
\item{density}{density of site}
\item{prev}{seropravalence}
}
}
\usage{
tortoise
}
\description{
Contains data regarding populations of Gopher Tortoises
}
\examples{
data(tortoise)
}
\keyword{datasets}
