% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_SE_CI.R
\name{SE_CI}
\alias{SE_CI}
\title{Final Project SE and CI}
\usage{
SE_CI(my_data = tortoise)
}
\arguments{
\item{my_data}{the data set you're using}
}
\value{
a list of the SE and CI of each estimate
}
\description{
Final Project SE and CI
}
\examples{
# load tortoise data
load("~/Gopher.Fit/data/tortoise.rda")

# compute SE and CI
SE_CI(tortoise)
}
